chore(gcli): address comments
